Fix broken Central Kurdish text in agentic UI on macOS
The agentic UI (apps/ui) is a separate renderer that never reflected the
active locale on the document element — it hardcoded lang="en" and only
loaded i18n data. As a result Kurdish text broke on macOS (the system
font lacks glyphs like ێ ۆ ڵ) and RTL locales rendered left-to-right.

Bundle the OFL-licensed Vazirmatn font (applied only for ckb) and set
document lang/dir from the resolved locale on bootstrap via a shared
applyLocale helper, replacing the duplicated per-entry translation loader.
