+++ apps/ui/src/main.hosted.tsx
@@ -1,26 +1,13 @@
-import { getLocaleData, isSupportedLocale } from '@studio/common/lib/locale';
-import { defaultI18n } from '@wordpress/i18n';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from '@/app';
 import { persistPromise } from '@/data/core';
 import { createHostedConnector } from '@/data/core/connectors/hosted';
-import type { Connector } from '@/data/core';
+import { applyLocale } from '@/lib/apply-locale';
 
 // Web entry point. Identical to `main.tsx` except it wires the HTTP/SSE hosted
 // connector instead of the Electron IPC connector, so the same React app runs
 // in a plain browser tab against the Studio hosted backend (`apps/hosted`).
-
-async function loadTranslations( connector: Connector ) {
-	const { locale } = await connector.getUserPreferences();
-	if ( ! locale || ! isSupportedLocale( locale ) ) {
-		return;
-	}
-	const translations = getLocaleData( locale )?.messages;
-	if ( translations ) {
-		defaultI18n.setLocaleData( translations );
-	}
-}
 
 function getDefaultApiBaseUrl(): string {
 	// Production builds are served by the Studio hosted backend itself, so the API is
@@ -34,7 +21,7 @@ async function bootstrap() {
 		apiBaseUrl: import.meta.env.VITE_STUDIO_API_URL ?? getDefaultApiBaseUrl(),
 	} );
 
-	await Promise.all( [ connector.init?.(), loadTranslations( connector ), persistPromise ] );
+	await Promise.all( [ connector.init?.(), applyLocale( connector ), persistPromise ] );
 
 	createRoot( document.getElementById( 'root' )! ).render(
 		<StrictMode>
