+++ apps/ui/src/main.tsx
@@ -1,27 +1,14 @@
-import { getLocaleData, isSupportedLocale } from '@studio/common/lib/locale';
-import { defaultI18n } from '@wordpress/i18n';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from '@/app';
 import { persistPromise } from '@/data/core';
 import { createIpcConnector } from '@/data/core/connectors/ipc';
-import type { Connector } from '@/data/core';
-
-async function loadTranslations( connector: Connector ) {
-	const { locale } = await connector.getUserPreferences();
-	if ( ! locale || ! isSupportedLocale( locale ) ) {
-		return;
-	}
-	const translations = getLocaleData( locale )?.messages;
-	if ( translations ) {
-		defaultI18n.setLocaleData( translations );
-	}
-}
+import { applyLocale } from '@/lib/apply-locale';
 
 async function bootstrap() {
 	const connector = createIpcConnector();
 
-	await Promise.all( [ connector.init?.(), loadTranslations( connector ), persistPromise ] );
+	await Promise.all( [ connector.init?.(), applyLocale( connector ), persistPromise ] );
 
 	createRoot( document.getElementById( 'root' )! ).render(
 		<StrictMode>
